Make tests conditional on safety profile. Move dynamic vwrite to new files.
