+++ dds/DCPS/XTypes/Utils.h
@@ -9,6 +9,7 @@
 #ifndef OPENDDS_SAFETY_PROFILE
 #  include <dds/DCPS/Serializer.h>
 #  include <dds/DCPS/Sample.h>
+#  include <dds/DCPS/DCPS_Utils.h>
 
 #  include <dds/DdsDynamicDataC.h>
 
@@ -242,6 +243,20 @@ inline DCPS::Sample::Extent nested(DCPS::Sample::Extent ext)
 // See description for ARRAY_TYPE in XTypes 1.3, page 139.
 DDS::ReturnCode_t flat_index(
   CORBA::ULong& flat_idx, const std::vector<CORBA::ULong>& idx_vec, const DDS::BoundSeq& dims);
+
+inline bool check_rc_from_get(DDS::ReturnCode_t rc, DDS::MemberId id, DDS::TypeKind tk,
+                              const char* fn_name, DCPS::LogLevel::Value log_thres = DCPS::LogLevel::Notice)
+{
+  if (rc != DDS::RETCODE_OK && rc != DDS::RETCODE_NO_DATA) {
+    if (DCPS::log_level >= log_thres) {
+      ACE_ERROR((DCPS::LogLevel::to_priority(log_thres), "(%P|t) %C: %C: Failed to get %C member ID %u: %C\n",
+                 DCPS::LogLevel::to_string(log_thres), fn_name,
+                 XTypes::typekind_to_string(tk), id, DCPS::retcode_to_string(rc)));
+    }
+    return false;
+  }
+  return true;
+}
 
 } // namespace XTypes
 } // namespace OpenDDS
